+++ node/__tests__/__utils__/mocks/adServer.ts
@@ -1,4 +1,7 @@
-import type { AdServerResponse } from '../../../typings/AdServer'
+import type {
+  AdServerResponse,
+  AdServerSponsoredBannersResponse,
+} from '../../../typings/AdServer'
 
 export const getSponsoredProductsResponse: AdServerResponse = {
   adRequestId: 'adRequestId',
@@ -12,6 +15,19 @@ export const getSponsoredProductsResponse: AdServerResponse = {
       options: {
         hideSponsoredBadge: true,
       },
+    },
+  ],
+}
+
+export const getSponsoredBannersResponse: AdServerSponsoredBannersResponse = {
+  adResponseId: 'adResponseBannerId',
+  banners: [
+    {
+      bannerImageId: 'bannerImageId',
+      imageUrl: 'https://example.com/banner.jpg',
+      targetUrl: 'https://example.com',
+      width: 970,
+      height: 250,
     },
   ],
 }