+++ node/resolvers/index.ts
@@ -1,7 +1,9 @@
+import { sponsoredBanners } from './sponsoredBanners'
 import { sponsoredProducts } from './sponsoredProducts'
 
 export const resolvers = {
   Query: {
     sponsoredProducts,
+    sponsoredBanners,
   },
 }