+++ node/typings/global.d.ts
@@ -91,4 +91,51 @@ declare global {
   type AdvertisementOptions = {
     hideSponsoredBadge?: boolean
   }
+
+  type Placement =
+    | 'top_search'
+    | 'middle_search'
+    | 'home_shelfpdp_shelf'
+    | 'search_shelf'
+    | 'cart_shelf'
+    | 'plp_shelf'
+    | 'autocomplete'
+    | 'homepage'
+
+  type AdUnit =
+    | 'billboard'
+    | 'smartphonebanner'
+    | 'leaderboard'
+    | 'superleaderboard'
+    | 'portrait'
+    | 'skyscraper'
+    | 'mediumrectangle'
+    | 'button120x60'
+    | 'mobilephoneinterstitial'
+    | 'featurephonesmallbanner'
+    | 'featurephonemediumbanner'
+    | 'featurephonelargebanner'
+
+  type Channel = 'website' | 'android' | 'ios' | 'msite' | 'whatsapp'
+
+  type SponsoredBannersParams = {
+    sponsoredCount: number
+    placement: Placement
+    adUnit: AdUnit
+    channel?: Channel
+    anonymousId?: string
+  }
+
+  type SponsoredBanner = {
+    adResponseId: string
+    advertisement: AdvertisementBanner
+  }
+
+  type AdvertisementBanner = {
+    bannerImageId: string
+    imageUrl: string
+    targetUrl: string
+    width: number
+    height: number
+  }
 }